Move default blueprint-id logic to install hanlder
Previously, this logic (of naming the blueprint id according to the
name of the containing directory of the main blueprint file)
was part of the `cfy blueprints` handler, inside the `upload` function.

I moved it to the `cfy install` handler since I didn't want this
change to affect current cli command, but rather just the new `install`
command. If we will find this behavior helpful, we could then make all
the other command act in a similar fashion.